Ajout des components auth, suppression des components inutiles du dossier register.
