+++ src/app/app.component.ts
@@ -2,13 +2,20 @@ import { Component } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { RouterOutlet } from '@angular/router';
 import {HomeComponent} from "./pages/home/home.component";
-import {LoginComponent} from "./pages/login/login.component";
-import {RegisterComponent} from "./pages/register/register.component";
+import {LoginComponent} from "./pages/auth/login/login.component";
+
 
 @Component({
   selector: 'app-root',
   standalone: true,
-  imports: [CommonModule, RouterOutlet, HomeComponent, LoginComponent, RegisterComponent],
+
+  imports: [
+    CommonModule,
+    RouterOutlet,
+    HomeComponent,
+    LoginComponent
+
+  ],
   templateUrl: './app.component.html',
   styleUrls: ['./app.component.scss']
 })
