+++ src/app/app.routes.ts
@@ -2,8 +2,7 @@ import { NgModule } from '@angular/core';
 import { Routes } from '@angular/router';
 import { RouterModule } from '@angular/router';
 import { HomeComponent} from "./pages/home/home.component";
-import { LoginComponent } from './pages/login/login.component';
-import { RegisterComponent } from './pages/register/register.component';
+import { LoginComponent } from './pages/auth/login/login.component';
 import {HeaderComponent} from "./components/header/header.component";
