--- src/app/pages/register/register.component.html
@@ -1,17 +1,0 @@
-<!-- register.component.html -->
-<form>
-  <label for="username">Identifiant</label>
-  <input type="text" id="username" name="username">
-
-  <label for="email">Adresse mail</label>
-  <input type="email" id="email" name="email">
-
-  <label for="password">Mot de passe</label>
-  <input type="password" id="password" name="password">
-
-  <label for="confirmPassword">Confirmez votre mot de passe</label>
-  <input type="password" id="confirmPassword" name="confirmPassword">
-
-  <button type="submit">Inscription</button>
-</form>
-
