--- src/app/pages/register/register.component.ts
@@ -1,12 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-  selector: 'app-register',
-  standalone: true,
-  imports: [],
-  templateUrl: './register.component.html',
-  styleUrl: './register.component.scss'
-})
-export class RegisterComponent {
-
-}
